+++ proofs/compiler/x86_instr_decl.v
@@ -128,7 +128,8 @@ Variant x86_op : Type :=
 | VPSHUFHW `(wsize)
 | VPSHUFLW `(wsize)
 | VPBLEND  `(velem) `(wsize)
-| VPBLENDVB `(wsize)
+| VPBLENDVB `(wsize) (* Deprecated: use BLENDV instead *)
+| BLENDV of velem & wsize (* vpblendvb, vblendvps, vblendvpd *)
 | VPACKUS  `(velem) `(wsize)
 | VPACKSS  `(velem) `(wsize)
 | VSHUFPS  `(wsize)
@@ -1559,14 +1560,32 @@ Definition Ox86_VPBLEND_instr :=
 
 Definition check_xmm_xmm_xmmm_xmm (_:wsize) := [:: [:: xmm; xmm; xmmm true; xmm]].
 
+(* TODO: remove in 2025.AA.0 *)
 Definition x86_VPBLENDVB sz (x y m: word sz) : tpl (w_ty sz) :=
   wpblendvb x y m.
 
+(* TODO: remove in 2025.AA.0 *)
 Definition Ox86_VPBLENDVB_instr :=
   (fun sz => mk_instr_safe
                (pp_sz "VPBLENDVB" sz) (w3_ty sz) (w_ty sz) [:: Eu 1; Eu 2; Eu 3] [:: Eu 0] MSB_CLEAR
                (@x86_VPBLENDVB sz) (check_xmm_xmm_xmmm_xmm sz) 4 (size_128_256 sz)
                (pp_name "vpblendvb" sz), ("VPBLENDVB"%string, prim_128_256 VPBLENDVB)).
+
+Definition x86_BLENDV ve sz (x y m: word sz) : tpl (w_ty sz) :=
+  blendv ve x y m.
+
+Definition Ox86_BLENDV_instr :=
+  (fun ve sz => mk_instr_safe
+                  (pp_ve_sz "BLENDV" ve sz) (w3_ty sz) (w_ty sz) [:: Ea 1; Eu 2; Ea 3] [:: Ea 0] MSB_CLEAR
+                  (@x86_BLENDV ve sz) (check_xmm_xmm_xmmm_xmm sz) 4 (size_128_256 sz)
+                  (pp_name match ve with
+                     | VE8 => "vpblendvb"
+                     | VE32 => "vblendvps"
+                     | VE64 => "vblendvpd"
+                     | _ => "<assert false>"
+                     end sz),
+     ("BLENDV"%string, primV_range [seq PVv ve sz | ve <- [:: VE8; VE32; VE64 ], sz <- [:: U128; U256 ]] BLENDV)
+  ).
 
 Definition SaturatedSignedToUnsigned (sz1 sz2:wsize) (w:word sz1) : word sz2 :=
   let i1 := wsigned w in
@@ -2146,6 +2165,7 @@ Definition x86_instr_desc o : instr_desc_t :=
   | VPUNPCKL sz sz'    => Ox86_VPUNPCKL_instr.1 sz sz'
   | VPBLEND ve sz      => Ox86_VPBLEND_instr.1 ve sz
   | VPBLENDVB sz       => Ox86_VPBLENDVB_instr.1 sz
+  | BLENDV ve sz       => Ox86_BLENDV_instr.1 ve sz
   | VPACKUS ve sz      => Ox86_VPACKUS_instr.1 ve sz
   | VPACKSS ve sz      => Ox86_VPACKSS_instr.1 ve sz
   | VPBROADCAST sz sz' => Ox86_VPBROADCAST_instr.1 sz sz'
@@ -2297,6 +2317,7 @@ Definition x86_prim_string :=
    Ox86_VPUNPCKL_instr.2;
    Ox86_VPBLEND_instr.2;
    Ox86_VPBLENDVB_instr.2;
+   Ox86_BLENDV_instr.2;
    Ox86_VPACKUS_instr.2;
    Ox86_VPACKSS_instr.2;
    Ox86_VPBROADCAST_instr.2;